Added the delete and send buttons. Selection is also implemented.
Decided to delete the name of the file and change it so that it appears
as a tool tip. There was a problem with the size of the image for the
file icon. For now it is going to stay this way because it is not that
large of a problem. The delete and send buttons have been added. The
delete has been implemented, the send has not yet been implemented
because how the files are being sent needs to be finished. The structure
is in place to send the files through the buttons, the backend is what
needs to be finished.

Next steps are:
-finish sending/receiving files (backend portion)
-work on modifying how files are sent/received
-find out more about max sizes for files and so on
-basically seeing how to speed it up or slow it down
-edit the gui to make it more beautiful
-Swrap: