File Transfer Protocol Completed
Files can now be sent and received between clients! Right now there is
no popup protection or anything like that. May want to include option to
download and so on. There are many possibilities to what is possible
now. The fact that the file transfering protocol is working is
fantastic!

Next steps are:
-work on modifying how files are sent/received
-finish the file sending/receive window (may make into two separate
windows)
-basically seeing how to speed it up or slow it down
-edit the gui to make it more beautiful
-Swrap:)