Two Users Can Connect and Disconnect
Two users can now connect and disconnect without any problems. The next
step is to make it so multiple users can connect and with each
connection they will share the new user and it will be added for
everyone on the server. Next step after that is for the implementation
of connection protocol for a password or so on. An "initial handshake".