Started Working on StartFrame, Connections are Sound
The connections work real well between users. Deciding whether or not to
make it so when connected to one user it will connect them to all the
users that they are connected to. However it is working real well. Need
to finish working out some little kinks of StartFrame. Also need it so
you can not connect to the same user twice, or add them twice if they
already exist. About ready to merge the this back to the original. Was
not sure if it was going to work but it did!

-Swrap